benchmarking/ttsbenchmarking/pyproject.toml
[project]
name = "ttsbenchmarking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "tts",
]

[project.scripts]
ttsbenchmarking = "ttsbenchmarking.app:main"

[build-system]
requires = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"

[tool.uv.sources]
tts = { workspace = true }
